requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
langchain==0.2.7
langchain-openai==0.1.14
python-dotenv==1.0.1
google-search-results==2.4.2
redis==5.0.1
